+++ src/platform/resource.h
@@ -130,6 +130,7 @@
 #define ID_PREF_WATCH_FILE 3165
 #define ID_PREF_PRINT_HEADER 3166
 #define ID_PREF_PRINT_FOOTER 3167
+#define ID_PREF_UPD_BUSY 3168
 
 // Static text controls
 #define IDC_STATIC -1